requirements.txt
azure-functions
sentence-transformers
